Fix up compile time problems - except for unused parameters
git-svn-id: https://mydns-ng.svn.sourceforge.net/svnroot/mydns-ng@177 fff143f6-a9c2-4b89-8f93-78cb2c0259b0

+++ mydns/branches/howard/pre-1.2.9/src/mydns/udp.c
@@ -139,7 +139,7 @@ write_udp_reply(TASK *t) {
     /* Err("%s: Send to (UDP) returned 0", desctask(t)); */
   }
 
-  if (rv != t->replylen) {
+  if (rv != (int)t->replylen) {
     /*
      * This should never ever happen as we have sent a partial packet over UDP
      */
